fix noise channel to not reset LFSR periodically
Previously, the LFSR was reset every time the channel was triggered,
which we perform once per tempo. This resulted in non-random predictable
audio, which is undesirable from the noise channel.

Now, the reset is tied to the global reset state, so the LFSR will be
constantly updating at all times, creating the 'random' noise in
a less predictable manner.